requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
python-dotenv
aiofiles
pydantic
web3
httpx
cryptography
greenlet
python-jose[cryptography]
passlib[bcrypt]
